requirements.txt
Django==1.8
Mezzanine==4.0.1
Pillow==2.9.0
beautifulsoup4==4.4.0
bleach==1.4.1
chardet==2.3.0
django-appconf==1.0.1
django-bootstrap3==5.3.1
django-ckeditor==4.5.1
django-contrib-comments==1.6.1
django-countries==3.3
django-forms-bootstrap==3.0.1
django-formtools==1.0
django-guardian==1.3
django-jsonfield==0.9.13
django-salesforce==0.6.1
django-toolbelt==0.0.1
django-user-accounts==1.0.1
eventlog==0.10.0
filebrowser-safe==0.4.0
future==0.15.0
grappelli-safe==0.4.1
gunicorn==19.3.0
html5lib==0.999999
oauthlib==1.0.3
pinax-phone-confirmation==0.1
pluggy==0.3.0
py==1.4.30
pytz==2014.10
requests==2.7.0
requests-oauthlib==0.5.0
selenium==2.47.1
simplejson==3.8.0
six==1.9.0
tox==2.1.1
tzlocal==1.2
virtualenv==13.1.2
wheel==0.24.0
